requirements.txt
Django>=1.5.4,<1.6
Jinja2==2.6
Markdown==2.2.1
South==0.7.6
boto==2.9.8
django-cms==2.4.1
django-mptt==0.5.2
django-compressor==1.3
django-polymorphic==0.4.0
django-reversion==1.6.6
django-jsonfield==0.8.10
django-phonenumber-field==0.2a3
phonenumbers==5.0b2
parsedatetime==0.8.7
pyparsing==1.5.6
statlib==1.1
python-magic==0.4.3
twilio==3.3.9
django-tastypie==0.9.15
python-dateutil==2.1
django-admin-tools==0.4.0
django-adminactions==0.0.4b3
cmsplugin-simple-markdown==1.2
django-kronos==0.3